Converting all c files into c++ files. Added a postfix and prefix operator overload for enumerators declared using enumerator_private.hpp.
https://tracker.physiomeproject.org/show_bug.cgi?id=3359

git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/branches/to_libzinc_and_beyond@9323 3c1deb5b-d424-0410-962d-aba41a686d42

+++ core/source/graphics/scene_viewer.h
@@ -62,7 +62,6 @@ translating and zooming with mouse button press and motion events.
 #include "graphics/light.h"
 #include "graphics/light_model.h"
 #include "graphics/scene.h"
-#include "interaction/interactive_tool.h"
 
 struct Graphics_buffer;
 #define Graphics_buffer_input Cmiss_scene_viewer_input
@@ -388,7 +387,7 @@ Sets the <scene_viewer> tumble angle.
 ==============================================================================*/
 
 int Scene_viewer_get_freespin_tumble_axis(struct Scene_viewer *scene_viewer,
-	float *tumble_axis);
+	double *tumble_axis);
 /*******************************************************************************
 LAST MODIFIED : 9 October 2003
 
@@ -398,7 +397,7 @@ about which the scene is turning relative to its lookat point.
 ==============================================================================*/
 
 int Scene_viewer_start_freespin(struct Scene_viewer *scene_viewer,
-	float *tumble_axis, double tumble_angle);
+	double *tumble_axis, double tumble_angle);
 /*******************************************************************************
 LAST MODIFIED : 10 September 2003
